v0.5.3

+++ interpreter/yasl_float.c
@@ -5,13 +5,13 @@
 #include <inttypes.h>
 
 char *float64_to_str(double d) {
-	int64_t size = snprintf(NULL, 0, "%f", d);
-	char *ptr = (char *)malloc(size + 1);
-	snprintf(ptr, size+1, "%f", d);
+	int size = snprintf(NULL, 0, "%f", d);
+	char *ptr = (char *)malloc((size_t)size+ 1);
+	snprintf(ptr, (size_t)size+1, "%f", d);
 	while (ptr[size - 1] == '0' && ptr[size - 2] != '.') {
 		size--;
 	}
 	ptr[size] = '\0';
-	ptr = (char *)realloc(ptr, size + 1);
+	ptr = (char *)realloc(ptr, (size_t)size + 1);
 	return ptr;
 }
